+++ grass-diary/src/components/Layout/Layout.tsx
@@ -1,8 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import Header from './Header';
 import Footer from './Footer';
-import SnackBar from './SnackBar';
-import { Toast, Modal } from '@components/index';
+import { Toast, Modal, SnackBar } from '@components/index';
 
 const Layout = () => {
   return (
